remove circular dep between core/scribbles and models/shop
