+++ app/core/scribbles.ts
@@ -2,9 +2,8 @@ import { getAltFormForPlayer, PkmsWithAltForms } from "../config"
 import type Player from "../models/colyseus-models/player"
 import type { Pokemon } from "../models/colyseus-models/pokemon"
 import PokemonFactory from "../models/pokemon-factory"
-import { getPokemonData } from "../models/precomputed/precomputed-pokemon-data"
+import { getPokemonData, getRegularsTier1 } from "../models/precomputed/precomputed-pokemon-data"
 import { PRECOMPUTED_POKEMONS_PER_RARITY } from "../models/precomputed/precomputed-rarity"
-import { getRegularsTier1 } from "../models/shop"
 import type GameState from "../rooms/states/game-state"
 import { Rarity } from "../types/enum/Game"
 import { Pkm, PkmRegionalVariants } from "../types/enum/Pokemon"
